+++ CHANGELOG.md
@@ -4,6 +4,11 @@
 Version 4.34-SNAPSHOT
 -------------
 
+CHANGED:
+
+- removed deprecated enum de.alpharogroup.file.checksum.Algorithm
+- removed deprecated enum de.alpharogroup.exception.ExceptionType
+- tagged several close methods as deprecated in class StreamExtensions since jdk7 the new feature try-with-resources statement
 
 Version 4.33
 -------------